requirements.txt
fastapi
uvicorn[standard]
python-multipart
qdrant-client
sqlalchemy
pymysql
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
openai
PyMuPDF
